+++ zarrs/src/array/chunk_grid/rectangular.rs
@@ -200,11 +200,11 @@ unsafe impl ChunkGridTraits for RectangularChunkGrid {
         self.chunks.len()
     }
 
-    fn array_shape(&self) -> &ArrayShape {
+    fn array_shape(&self) -> &[u64] {
         &self.array_shape
     }
 
-    fn grid_shape(&self) -> &ArrayShape {
+    fn grid_shape(&self) -> &[u64] {
         &self.grid_shape
     }