+++ zarrs_chunk_grid/CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `ChunkGridTraitsIterators::iter_chunk_indices_and_subsets()`
 - **Breaking**: Change `Item` associated type of `[Par]Indices[Into]Iterator` from `ArrayIndices` to `ArrayIndicesTinyVec`
 - **Breaking**: Change `Item` associated type of `[Par]ContiguousIndices[Into]Iterator` from `(ArrayIndices, u64)` to `(ArrayIndicesTinyVec, u64)`
+- **Breaking**: Change return type `ChunkGridTraits::{array_shape,grid_shape}()` to `&[u64]` instead of `&ArrayShape` 
 
 
 [unreleased]: https://github.com/zarrs/zarrs/compare/zarrs_chunk_grid-v0.1.0...HEAD